Emit RIME event on successful device decryption
Emit RIME event when a device is successfully decrypted so
that GraphQL-subscribed clients receive the updated state.
